+++ ObitSystem/ObitSD/python/Obit_wrap.c
@@ -2880,7 +2880,7 @@ extern PyObject *ImageDescGetDict(ObitImageDesc* in) {
   PyObject *outDict = PyDict_New();
   PyObject *list1, *list2, *list3, *list4, *list5, *list6;
   PyObject *value;
-  Py_ssize_t doh;
+  int doh;
   int i, pos = 0;
 
   PyDict_SetItemString(outDict, "name",    PyString_InternFromString(in->name));
@@ -2941,7 +2941,7 @@ extern PyObject *ImageDescGetDict(ObitImageDesc* in) {
   PyDict_SetItemString(outDict, "jlocf",    PyInt_FromLong((long)in->jlocf));
 
   /* Discard references to newly created objects. */
-  doh = (Py_ssize_t)pos;
+  doh = pos;
   while (PyDict_Next(outDict, &doh, NULL, &value))
     Py_DECREF(value);
 
@@ -7977,7 +7977,7 @@ extern PyObject *TableDescGetDict(ObitTableDesc* in) {
   PyObject *outDict = PyDict_New();
   PyObject *list, *value;
   gchar *ctemp;
-  Py_ssize_t doh;
+  int doh;
   int i, pos = 0;
 
   /* test validity */
@@ -8033,7 +8033,7 @@ extern PyObject *TableDescGetDict(ObitTableDesc* in) {
   PyDict_SetItemString(outDict, "FieldUnit", list);
 
   /* Discard references to newly created objects. */
-  doh = (Py_ssize_t)pos;
+  doh = pos;
   while (PyDict_Next(outDict, &doh, NULL, &value))
     Py_DECREF(value);
 
@@ -11568,7 +11568,7 @@ extern PyObject *UVDescGetDict(ObitUVDesc* in) {
   PyObject *outDict = PyDict_New();
   PyObject *list1, *list2, *list3, *list4, *list5, *list6, *list7;
   PyObject *value;
-  Py_ssize_t doh;
+  int doh;
   int i, pos = 0;
 
   PyDict_SetItemString(outDict, "name",    PyString_InternFromString(in->name));
@@ -11645,7 +11645,7 @@ extern PyObject *UVDescGetDict(ObitUVDesc* in) {
   PyDict_SetItemString(outDict, "incif",    PyInt_FromLong((long)in->incif));
  
   /* Discard references to newly created objects. */
-  doh = (Py_ssize_t)pos;
+  doh = pos;
   while (PyDict_Next(outDict, &doh, NULL, &value))
     Py_DECREF(value);
